Output byte extents are computed in u64
Both call sites computed an output byte extent in usize and passed it to with_output_byte_range, which takes u64, so vyre-libs-math did not compile once a caller enabled the math features. prefix_scan additionally fell back to usize::MAX and then saturated the multiply, which produces a wrong extent instead of an error. Both compute in u64 now, and the scan extent is exact because a u32 element count times four always fits.

+++ vyre-libs-math/src/math/linalg/matmul_tiled/program.rs
@@ -172,8 +172,8 @@ pub(crate) fn build_matmul_tiled_program(
             name: out.to_string(),
             shape: vec![m, n],
         })?;
-    let logical_output_bytes = (logical_out_count as usize)
-        .checked_mul(element_size)
+    let logical_output_bytes = u64::from(logical_out_count)
+        .checked_mul(element_size as u64)
         .ok_or_else(|| TensorRefError::ElementCountOverflow {
             name: out.to_string(),
             shape: vec![m, n],
